requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
transformers==4.38.2
Pillow==10.2.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.0.3
numpy==1.26.4
torch==2.2.1
gdown==5.2.0

#torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu